requirements.txt
fastapi
uvicorn
jinja2
python-multipart
ai-edge-litert
librosa
numpy
pillow
scikit-learn
matplotlib
